flask_backend/requirements.txt
flask
motor
requests
asyncio
python-dotenv
flask[async]